#14 refactor: 패키지 구조 변경
도메인 별로 패키지를 변경합니다.
