+++ src/main/java/com/fourdays/core/domain/url/model/entity/exception/InvalidUrlKeyException.java
@@ -1,4 +1,4 @@
-package com.fourdays.core.model.domain.url.entity.exception;
+package com.fourdays.core.domain.url.model.entity.exception;
 
 public class InvalidUrlKeyException extends InvalidException {
