+++ src/main/java/com/fourdays/core/domain/url/model/repository/UrlRepositoryMemory.java
@@ -1,7 +1,7 @@
-package com.fourdays.core.model.domain.url.repository;
+package com.fourdays.core.domain.url.model.repository;
 
-import com.fourdays.core.model.domain.url.entity.URL;
-import com.fourdays.core.model.domain.url.entity.exception.InvalidUrlKeyException;
+import com.fourdays.core.domain.url.model.entity.URL;
+import com.fourdays.core.domain.url.model.entity.exception.InvalidUrlKeyException;
 import org.springframework.util.StringUtils;
 
 import java.util.HashMap;
