+++ src/test/java/com/fourdays/core/domain/url/model/repository/UrlRepositoryMemoryTest.java
@@ -1,7 +1,7 @@
-package com.fourdays.core.model.domain.url.repository;
+package com.fourdays.core.domain.url.model.repository;
 
-import com.fourdays.core.model.domain.url.entity.URL;
-import com.fourdays.core.model.domain.url.entity.exception.InvalidUrlKeyException;
+import com.fourdays.core.domain.url.model.entity.URL;
+import com.fourdays.core.domain.url.model.entity.exception.InvalidUrlKeyException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
